adding -Wconversion
git-svn-id: https://svn.code.sf.net/p/lass/code/lass/trunk@1100 e9412239-ebb7-49aa-8f4a-f9aa784616d7

+++ test_suite/bar.cpp
@@ -188,7 +188,7 @@ namespace lass
 
 		void  Bar::testAutomaticFunctionExport( int iA, float iB )
 		{
-			LASS_COUT << "C++ testAutomaticFunctionExport : " << iA + iB << "\n";
+			LASS_COUT << "C++ testAutomaticFunctionExport : " << static_cast<float>(iA) + iB << "\n";
 		}
 
 		void Bar::complexArguments( const std::string& iA )
